K1: partial revert of KT-49507
Note: we prefer base class field against derived class property
only in case when property is not visible.
In this case we keep the fix of KT-49507 intact.
Otherwise, we have to revert it, because in future we want to choose
the property in this case, and we don't want to implement semantic
breaking change.

See also KT-54393
KT-34943 is kept fixed